Monkey patch for date sorting
I removed past dates. While doing this, I found the denver date to be
before the rest. This isn't good. So, I looked at the code. It does a
for loop and creates the entry before moving on to the next one. This
means that things are listed in reverse chronological order. While I
don't know enough about Ruby to fix this elegantly, I've renamed all
the files to be "2012-01-XX" and when a new entry is added, the XX for
the day should be incremented down. This will ensure the posts are done
in the proper order. When "01" is reached for XX, it will need to be
manually reordered again. I'll keep an eye out for an elegant fix. I'm
certain one can be found.