Removed old class postings
Also modified the index files of each class type to display a
message when no classes were present based on if there was
even 1 tag present.

Checking for the tag also allowed stub files in the in-person
class section. This stub doesnt seem to be needed in web-based
because most of the files are duplicates anyways.